requirements.txt
langchain==0.2.16
langchain-core>=0.2.38,<0.3
langchain-anthropic==0.1.23
langchain-groq==0.1.9
langchain-community==0.2.16
anthropic>=0.30.0
groq>=0.9.0
httpx==0.27.2
requests==2.32.3
python-dotenv==1.0.1
pydantic==2.8.2
rich==13.8.1
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
aiofiles==24.1.0
sentence-transformers==3.0.1
numpy>=1.24.0
playwright>=1.44.0
